+++ packages/presets/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2023-08-09
+
 ### Fixed
 
 - Resolved an issue with bundling the `@park-ui/presets` package.
@@ -45,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Release of https://park-ui.com 🎉
 
-[unreleased]: https://github.com/cschroeter/park-ui/compare/@park-ui/presets@0.3.0...HEAD
+[unreleased]: https://github.com/cschroeter/park-ui/compare/@park-ui/presets@0.3.1...HEAD
+[0.3.1]: https://github.com/cschroeter/park-ui/compare/@park-ui/presets@0.3.0...@park-ui/presets@0.3.1
 [0.3.0]: https://github.com/cschroeter/park-ui/compare/@park-ui/presets@0.2.0...@park-ui/presets@0.3.0
 [0.2.0]: https://github.com/cschroeter/park-ui/compare/@park-ui/presets@0.1.0...@park-ui/presets@0.2.0
 [0.1.0]: https://github.com/cschroeter/park-ui/releases/tag/@ark-ui/react@0.1.0